move auth code to central place, refactor general new-Session code
